requirements.txt
pandas
tensorflow
keras
numpy
mysql-connector-python
scikit-learn
dataclasses
dill
flask
nltk
regex
scikeras[tensorflow]
-e .